Add pydeck types
