+++ bindings/pydeck/pydeck/types/__init__.py
@@ -1,0 +1,2 @@
+from .image import Image  # noqa
+from .string import String  # noqa
